tighten compliance requirement for CoC
The CoC has components that apply to corporate users, and in particular in ways that may clash with entire business models ("Do not ... Use ml5.js to build tools of mass surveillance"). "all practical steps", while perhaps appropriate for a notice requirement, is perhaps an out clause for use that would otherwise require a change in business model (impractical for most businesses!) 

This PR tightens the language, in light of that concern.

Note that accepting this PR may make the conversation proposed in issue #2 (already unlikely to be easy) more complex.